Refactor UDP code to make room for TCP feature
This is a breaking change. Clients will need to use gelf.NewUDPWriter
instead of gelf.NewWriter. The type returned will also change from
*gelf.Writer to *gelf.UDPWriter, or gelf.Writer for the interface.
